analyzer/requirements.txt
nats-py
python-dotenv
dataclasses-json
dacite
aiohttp
requests
